The TEA might be implemented correctly.
